+++ packages/vm/src/wasm_backend/mod.rs
@@ -8,4 +8,4 @@ mod metering;
 pub use engine::make_compiler_config;
 
 pub use compile::compile;
-pub use engine::{make_compiling_engine, make_runtime_engine};
+pub use engine::{make_compiling_engine, make_runtime_engine, COST_FUNCTION_HASH};